pyproject.toml
[tool.poetry]
name = "News sosh bot"
version = "0.1.0"
description = ""
authors = ["CostGem"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.2.0"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.24"
asyncpg = "^0.29.0"
fluentogram = "^1.1.6"
alembic = "^1.13.1"
greenlet = "^3.0.3"
aiolimiter = "^1.1.0"
cachetools = "^5.3.2"
dataclasses-json = "^0.6.4"
apscheduler = "3.10.4"
apscheduler-di = "0.1.0"
dependency-injector = "^4.42.0"
environ-config = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
